01-docker/pyproject.toml
[project]
name = "01-docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "pgcli>=4.3.0",
    "psycopg2-binary>=2.9.11",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.43",
]
